Move job list components to queues folder
